Rename update channel
