+++ packages/server-wallet/src/handlers/__test__/update-channel.test.ts
@@ -1,0 +1,6 @@
+import {updateChannel} from '../update-channel';
+import {updateChannelFixture, channelStateFixture} from '../fixtures/update-channel';
+
+test('validUpdate', () => {
+  updateChannel(updateChannelFixture(), channelStateFixture());
+});
