--- packages/server-wallet/src/handlers/__test__/updateChannel.test.ts
@@ -1,6 +1,0 @@
-import {updateChannel} from '../updateChannel';
-import {updateChannelFixture, channelStateFixture} from '../fixtures/updateChannel';
-
-test('validUpdate', () => {
-  updateChannel(updateChannelFixture(), channelStateFixture());
-});
